ui реализован

+++ app/src/main/java/com/example/ip_test_task/domain/entity/ShopItem.kt
@@ -1,9 +1,11 @@
 package com.example.ip_test_task.domain.entity
 
+import android.icu.util.Calendar
+
 data class ShopItem(
     val id: Int,
     val name: String,
-    val time: Int,
+    val time: Long,
     val tags: List<String>,
     val amount: Int
 )
